requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
Deprecated==1.3.1
distro==1.9.0
frozenlist==1.8.0
get-annotations==0.1.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-tools==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.14.0
magic-filter==1.0.12
maxapi==0.9.18
multidict==6.7.1
openai==2.31.0
packaging==26.0
propcache==0.4.1
protobuf==6.33.6
puremagic==1.30
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==2.1.2
yandex-ai-studio-sdk==0.20.1
yandex-cloud-ml-sdk==0.19.1
yandexcloud==0.386.0
yarl==1.23.0
ydb==3.28.0
sqlalchemy==2.0.29